Added condition to check file exists before removing it

+++ bin/prepare-heroku-deployment.sh
@@ -1,3 +1,12 @@
+#!/bin/bash
+#
 # remove these files so that Heroku doesn't detect Node and try to do a Node build
-rm package.json
-rm package-lock.json
+#
+
+if [ -f package.json ] ; then
+	rm package.json
+fi
+
+if [ -f package-lock.json ] ; then
+	rm package-lock.json
+fi
